Fix bugs

+++ database/migrations/2020_02_06_145406_create_posts_table.php
@@ -20,7 +20,7 @@ class CreatePostsTable extends Migration
             $table->index('author_id');
             $table->string('header');
             $table->text('content');
-            $table->bigInteger('rating');
+            $table->integer('rating');
             $table->index('rating');
             $table->timestamps();
         });
